refactor: generate barrel & use proper import
