+++ package.json
@@ -16,6 +16,7 @@
     "dev": "tiddlywiki-plugin-dev dev",
     "build": "npm run clean && tiddlywiki-plugin-dev build",
     "publish": "npm run clean && tiddlywiki-plugin-dev publish",
+    "generate-barrels": "barrelsby --delete --location below --singleQuotes --directory ./src/slate-write/editor --exclude .+\\.d\\.ts && dprint fmt **/*/index.ts",
     "reset": "rimraf ./**/node_modules",
     "clean": "rimraf dist",
     "prepare": "husky install",
@@ -46,6 +47,7 @@
     "@types/react": "^18.2.14",
     "@types/react-dom": "^18.2.6",
     "@types/styled-components": "^5.1.26",
+    "barrelsby": "^2.8.0",
     "beautiful-react-hooks": "3.11.2",
     "dprint": "^0.37.1",
     "eslint-config-tidgi": "1.1.1",
@@ -106,7 +108,6 @@
     "@udecode/plate-styled-components": "21.3.2",
     "@udecode/plate-table": "21.3.4",
     "@udecode/plate-trailing-block": "21.3.2",
-    "@udecode/plate-ui": "21.4.0",
     "@udecode/plate-ui-block-quote": "21.3.2",
     "@udecode/plate-ui-dnd": "21.3.2",
     "@udecode/plate-ui-link": "21.4.0",
