+++ src/slate-write/editor/components/drag-and-drop/withStyledDraggables.tsx
@@ -7,8 +7,8 @@ import { ELEMENT_PARAGRAPH } from '@udecode/plate-paragraph';
 import { ELEMENT_H1, ELEMENT_H2, ELEMENT_H3, ELEMENT_H4, ELEMENT_H5, ELEMENT_H6 } from '@udecode/plate-heading';
 
 import type { PlatePluginComponent } from '@udecode/plate-core';
-import type { DefaultPlatePluginKey } from '@udecode/plate-ui';
 import { withPlateDraggables } from '@udecode/plate-ui-dnd';
+import { DefaultPlatePluginKey } from '../../config/DefaultPlatePluginKey';
 import { ELEMENT_CODE_BLOCK } from '../../plugins/codeblock/constants';
 import { ELEMENT_MACRO } from '../../plugins/macro';
 import { ELEMENT_WIDGET } from '../../plugins/widget';
