+++ src/slate-write/editor/config/DefaultPlatePluginKey.ts
@@ -1,0 +1,47 @@
+import { MARK_BOLD, MARK_CODE, MARK_ITALIC, MARK_STRIKETHROUGH, MARK_SUBSCRIPT, MARK_SUPERSCRIPT, MARK_UNDERLINE } from '@udecode/plate-basic-marks';
+import { ELEMENT_BLOCKQUOTE } from '@udecode/plate-block-quote';
+import { ELEMENT_CODE_BLOCK, ELEMENT_CODE_LINE } from '@udecode/plate-code-block';
+import { ELEMENT_H1, ELEMENT_H2, ELEMENT_H3, ELEMENT_H4, ELEMENT_H5, ELEMENT_H6 } from '@udecode/plate-heading';
+import { MARK_HIGHLIGHT } from '@udecode/plate-highlight';
+import { MARK_KBD } from '@udecode/plate-kbd';
+import { ELEMENT_LINK } from '@udecode/plate-link';
+import { ELEMENT_LI, ELEMENT_OL, ELEMENT_TODO_LI, ELEMENT_UL } from '@udecode/plate-list';
+// import { ELEMENT_IMAGE, ELEMENT_MEDIA_EMBED } from '@udecode/plate-media';
+import { ELEMENT_MENTION, ELEMENT_MENTION_INPUT } from '@udecode/plate-mention';
+import { ELEMENT_PARAGRAPH } from '@udecode/plate-paragraph';
+import { ELEMENT_TABLE, ELEMENT_TD, ELEMENT_TH, ELEMENT_TR } from '@udecode/plate-table';
+import { ELEMENT_WIDGET } from '../plugins';
+
+export type DefaultPlatePluginKey =
+  // modify plate's packages/ui/plate/src/types/DefaultPlatePluginKey.ts to add this
+  | typeof ELEMENT_WIDGET
+  | typeof ELEMENT_BLOCKQUOTE
+  | typeof ELEMENT_CODE_BLOCK
+  | typeof ELEMENT_CODE_LINE
+  | typeof ELEMENT_H1
+  | typeof ELEMENT_H2
+  | typeof ELEMENT_H3
+  | typeof ELEMENT_H4
+  | typeof ELEMENT_H5
+  | typeof ELEMENT_H6
+  | typeof ELEMENT_LI
+  | typeof ELEMENT_LINK
+  | typeof ELEMENT_MENTION
+  | typeof ELEMENT_MENTION_INPUT
+  | typeof ELEMENT_OL
+  | typeof ELEMENT_PARAGRAPH
+  | typeof ELEMENT_TABLE
+  | typeof ELEMENT_TD
+  | typeof ELEMENT_TH
+  | typeof ELEMENT_TODO_LI
+  | typeof ELEMENT_TR
+  | typeof ELEMENT_UL
+  | typeof MARK_BOLD
+  | typeof MARK_CODE
+  | typeof MARK_HIGHLIGHT
+  | typeof MARK_ITALIC
+  | typeof MARK_KBD
+  | typeof MARK_STRIKETHROUGH
+  | typeof MARK_SUBSCRIPT
+  | typeof MARK_SUPERSCRIPT
+  | typeof MARK_UNDERLINE;
