+++ src/slate-write/editor/config/index.ts
@@ -1,0 +1,9 @@
+/**
+ * @file Automatically generated by barrelsby.
+ */
+
+export * from './autoformat/index';
+export * from './config';
+export * from './globalStyle';
+export * from './plugins';
+export * from './withStyledPlaceHolders';
