+++ src/slate-write/editor/config/withStyledPlaceHolders.ts
@@ -2,8 +2,8 @@
 /* eslint-disable @typescript-eslint/explicit-function-return-type */
 import type { PlatePluginComponent } from '@udecode/plate-core';
 import { ELEMENT_PARAGRAPH } from '@udecode/plate-paragraph';
-import type { DefaultPlatePluginKey } from '@udecode/plate-ui';
 import { withPlaceholders } from '@udecode/plate-ui-placeholder';
+import { DefaultPlatePluginKey } from './DefaultPlatePluginKey';
 
 export const withStyledPlaceHolders = (components: any): Record<DefaultPlatePluginKey, PlatePluginComponent<any>> =>
   withPlaceholders(components, [
