--- src/slate-write/editor/plugins/autoComplete/index.ts
@@ -1,7 +1,0 @@
-/**
- * @file Automatically generated by barrelsby.
- */
-
-export * from './createAutoCompletePlugin';
-export * from './types';
-export * from './withAutoComplete';
