--- src/slate-write/editor/plugins/autoComplete/queries/index.ts
@@ -1,7 +1,0 @@
-/**
- * @file Automatically generated by barrelsby.
- */
-
-export * from './findAutoCompleteInput';
-export * from './isNodeAutoCompleteInput';
-export * from './isSelectionInAutoCompleteInput';
