--- src/slate-write/editor/plugins/autoComplete/transforms/index.ts
@@ -1,5 +1,0 @@
-/**
- * @file Automatically generated by barrelsby.
- */
-
-export * from './removeAutoCompleteInput';
