--- src/slate-write/editor/plugins/codeblock/CodeBlockElement/index.ts
@@ -1,6 +1,0 @@
-/**
- * @file Automatically generated by barrelsby.
- */
-
-export * from './CodeBlockElement';
-export * from './CodeBlockSelectElement';
