--- src/slate-write/editor/plugins/codeblock/index.ts
@@ -1,7 +1,0 @@
-/**
- * @file Automatically generated by barrelsby.
- */
-
-export * from './CodeBlockElement';
-export * from './constants';
-export * from './createCodeBlockPlugin';
