--- src/slate-write/editor/plugins/codeblock/options/index.ts
@@ -1,5 +1,0 @@
-/**
- * @file Automatically generated by barrelsby.
- */
-
-export * from './getCodeLineType';
