+++ src/slate-write/editor/plugins/comboBox/getAutoCompleteOnSelectItem.ts
@@ -9,7 +9,7 @@ import { deleteText, insertNodes, insertText, isEndPoint, TNodeProps, TText, Val
 import { getBlockAbove } from '@udecode/slate-utils';
 import { replaceCurrentBlockWithParseResult } from '../advancedBlock/replaceCurrentBlockWithParseResult';
 import { ELEMENT_AUTO_COMPLETE } from '../autoComplete/createAutoCompletePlugin';
-import { removeAutoCompleteInputFromCurrentSelection } from '../autoComplete/transforms';
+import { removeAutoCompleteInputFromCurrentSelection } from '../autoComplete/transforms/removeAutoCompleteInput';
 import { AutoCompletePlugin, TAutoCompleteElement } from '../autoComplete/types';
 import { ComboboxOnSelectItem } from './Combobox.types';
 import { useAutoCompletePluginStore } from './store';
